requirements.txt
numpy                       2.2.6
albumentations              2.0.8
matplotlib                  3.10.6
pillow                      11.3.0
pytorch-lightning           2.5.5
segmentation_models_pytorch 0.5.0
torch                       2.8.0
torchmetrics                1.8.2
torchvision                 0.23.0
tqdm                        4.67.1